Fixed model structure + other small changings
