+++ todo_list/.env.example
@@ -1,0 +1,13 @@
+# .env
+
+# Django settings
+SECRET_KEY=
+
+# Database settings
+DB_NAME=
+DB_USER=
+DB_PASSWORD=
+DB_HOST=
+DB_PORT=
+
+# Other settings...
