+++ todo_list/todo_list_app/models.py
@@ -1,22 +1,5 @@
-from django.db import models
-from django.contrib.auth.models import User
+
 # Create your models here.
 
-
-class Category(models.Model):
-    name = models.CharField(max_length=100)
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return self.name
-
-
-class Task(models.Model):
-    title = models.CharField(max_length=100)
-    description = models.TextField(blank=True,null=True)
-    due_date = models.DateField()
-    completed = models.BooleanField(default=False)
-    category = models.ForeignKey(Category, on_delete=models.CASCADE)
-
-    def __str__(self):
-        return self.title
+from .models import Category
+from .models import Task
