+++ todo_list/todo_list_app/models/__init__.py
@@ -1,0 +1,4 @@
+from .categories import Category
+from .tasks import Task
+
+__all__ = ['Category', 'Task']
