+++ todo_list/todo_list_app/models/categories.py
@@ -1,0 +1,10 @@
+from django.db import models
+from django.contrib.auth.models import User
+
+
+class Category(models.Model):
+    name = models.CharField(max_length=100)
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
+
+    def __str__(self):
+        return self.name
