+++ todo_list/todo_list_app/models/tasks.py
@@ -1,0 +1,14 @@
+from django.db import models
+from .categories import Category
+
+
+class Task(models.Model):
+    title = models.CharField(max_length=100)
+    description = models.TextField(blank=True, null=True)
+    due_date = models.DateField()
+    completed = models.BooleanField(default=False)
+    category = models.ForeignKey(Category, on_delete=models.CASCADE)
+
+    def __str__(self):
+        return self.title
+
